requirements.txt
groq>=0.9.0
sounddevice>=0.4.6
numpy>=1.24.0
keyboard>=0.13.5
pyperclip>=1.8.2
pystray>=0.19.5
Pillow>=10.0.0
python-dotenv>=1.0.0
scipy>=1.11.0
